% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_date}
\alias{is_date}
\title{Is Date?
Internal Function to check that an object
is a date}
\usage{
is_date(x)
}
\arguments{
\item{x}{vector, ideally of dates}
}
\description{
Is Date?
Internal Function to check that an object
is a date
}
