% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_tile_plot.R
\name{make_and_export_year_tile_png}
\alias{make_and_export_year_tile_png}
\title{Make and Export Year Tile PNG
creates the plot and exports to file
with the correct dimensions}
\usage{
make_and_export_year_tile_png(
  df,
  fill_palette = "Greens",
  path = "mygraph.png"
)
}
\arguments{
\item{df}{data.frame like output of make_year_df}

\item{fill_palette}{string brewer palette. defaults to 'Greens'}

\item{path}{string path to save. defaults to 'mygraph.png'}
}
\value{
saved png
}
\description{
Make and Export Year Tile PNG
creates the plot and exports to file
with the correct dimensions
}
\examples{
/dontrun{
fake <- data.frame(
date = seq(as.Date("2023-01-01"), as.Date("2023-12-31"), by = "1 day"),
 n = sample(x = 0:10, size = 365, replace = TRUE)
)
make_and_export_year_tile_png(df = make_year_df(year = 2023, fake))
}
}
