% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_year_df.R
\name{make_empty_year_df}
\alias{make_empty_year_df}
\title{Make Empty Year DF}
\usage{
make_empty_year_df(year = 2024)
}
\arguments{
\item{year}{numeric defaults to 2024}
}
\value{
data.frame with columns needed for year_tile_plot
}
\description{
Make Empty Year DF
}
\examples{
tmp <- make_empty_year_df()

# can submit a year
tmp <- make_empty_year_df('2012')
}
