% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_year_df.R
\name{make_year_df}
\alias{make_year_df}
\title{Make Year DF}
\usage{
make_year_df(year = 2024, day_df)
}
\arguments{
\item{year}{numeric defaults to 2024}

\item{day_df}{data.frame expected to have cols date and n}
}
\value{
data.frame with columns needed for year_tile_plot
}
\description{
Make Year DF
}
\examples{
fake <- data.frame(
date = seq(as.Date("2023-01-01"), as.Date("2023-12-31"), by = "1 day"),
 n = sample(x = 0:10, size = 365, replace = TRUE)
)
make_year_df(year = 2023, day_df = fake)
}
