% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_tile_plot.R
\name{make_year_tile_plot}
\alias{make_year_tile_plot}
\title{Make Year Tile Plot
creates a yearly graph for any measure showing
the distribution of effort using any brewer palette
inspired by github contributions plot
code modified from https://restateinsight.com/posts/general-posts/2024-12-github-contributions-plot/}
\usage{
make_year_tile_plot(
  year_df,
  fill_palette = c("Blues", "BuGn", "BuPu", "GnBu", "Greens", "Greys", "Oranges", "OrRd",
    "PuBu", "PuBuGn", "PuRd", "Purples", "RdPu", "Reds", "YlGn", "YlGnBu", "YlOrBr",
    "YlOrRd")
)
}
\arguments{
\item{year_df}{df with day and n}

\item{fill_palette}{string brewer palette; defaults to Blues but to make it look like github contributions, you would use 'Greens'}
}
\value{
ggplot object
}
\description{
Make Year Tile Plot
creates a yearly graph for any measure showing
the distribution of effort using any brewer palette
inspired by github contributions plot
code modified from https://restateinsight.com/posts/general-posts/2024-12-github-contributions-plot/
}
\examples{
/dontrun{
fake <- data.frame(
date = seq(as.Date("2023-01-01"), as.Date("2023-12-31"), by = "1 day"),
 n = sample(x = 0:10, size = 365, replace = TRUE)
)
# will print a distorted year tile plot (too)
make_year_tile_png(df = make_year_df(year = 2023, fake))
# should have dimensions width = 5xheight to look correct
}
}
